add new file test_processing.py in packet tests
