+++ README.md
@@ -74,14 +74,31 @@ def sort_by_date(list_diction: Iterable[dict])-> Iterable[dict]:
     return sorted_list_data
 ```
 ### Пример входных данных для проверки 
-1) функция filter_by_state
+1) функция get_mask_card_number
+
+    7000 79** **** 6361
+
+2) функция get_mask_account
+
+    **4305
+
+3) функция mask_account_card
+    
+    Visa Platinum 7000 79** **** 6361
+    Счет **4305
+
+4) функция get_date
+    11.03.2024
+
+5) функция filter_by_state
 
 [{'id': 41428829, 'state': 'EXECUTED', 'date': '2019-07-03T18:35:29.512364'}, {'id': 939719570, 'state': 'EXECUTED', 'date': '2018-06-30T02:08:58.425572'}, {'id': 594226727, 'state': 'CANCELED', 'date': '2018-09-12T21:27:25.241689'}, {'id': 615064591, 'state': 'CANCELED', 'date': '2018-10-14T08:21:33.419441'}]
-2) функция filter_by_state
+
+6) функция filter_by_state
 
 [{'id': 41428829, 'state': 'EXECUTED', 'date': '2019-07-03T18:35:29.512364'}, {'id': 939719570, 'state': 'EXECUTED', 'date': '2018-06-30T02:08:58.425572'}, {'id': 594226727, 'state': 'CANCELED', 'date': '2018-09-12T21:27:25.241689'}, {'id': 615064591, 'state': 'CANCELED', 'date': '2018-10-14T08:21:33.419441'}]
 
 ## Тесты
-1) К функциям get_mask_card_number и get_mask_account тесты написаны в тестовом файле test_masks.py, функциональный код покрыт тестами на 93%.
+1) К функциям get_mask_card_number и get_mask_account тесты написаны в тестовом файле test_masks.py, функциональный код покрыт тестами на 96%.
 2) К функциям mask_account_card и get_date тесты написаны в тестовом файле test_widget.py, функциональный код покрыт тестами на 100%.
-3) К функциям filter_by_state и sort_by_date тесты написаны в тестовом файле test_
+3) К функциям filter_by_state и sort_by_date тесты написаны в тестовом файле test_processing, функциональный код покрыт тестами на 100%.
