+++ main.py
@@ -1,6 +1,6 @@
 from typing import Union
 
-from src.masks import get_mask_card_number, get_mask_account
+from src.masks import get_mask_account, get_mask_card_number
 from src.processing import filter_by_state, sort_by_date
 from src.widget import mask_account_card, get_date
 
@@ -12,6 +12,12 @@ def print_hi(name: Union[str]) -> Union[str]:
 
 if __name__ == "__main__":
     print_hi("PyCharm")
+
+print(get_mask_card_number(7000792289606361))
+print(get_mask_account(70007922896063616541))
+print(mask_account_card("Visa Platinum 7000792289606361"))
+print(mask_account_card("Счет 73654108430135874305"))
+print(get_date("2024-03-11T02:26:18.671407"))
 
 print(
     filter_by_state(
@@ -35,6 +41,3 @@ print(
     )
 )
 
-print(mask_account_card('Visa Platinum 8990922113665229'))
-print(mask_account_card('Visa Gold 5999414228426353'))
-print(get_date(""))
